Updates to add Admin pages and login and final features

+++ server/models/admin.js
@@ -1,13 +1,17 @@
 const db = require('../db');
 const { v4: uuidv4 } = require('uuid');
 const sqlstring = require('sqlstring');
+const bcrypt = require('bcryptjs');
 
 module.exports = {
-  verify: function (data, callback) {
+  verify: function (uuid, callback) {
     db.connection.connect();
-    db.connection.query(`SELECT * FROM admin WHERE uuid = '${data.uuid}'`, null, (err, results) => {
+    db.connection.query(`SELECT * FROM admin WHERE uuid = '${uuid}'`, null, (err, results) => {
+      console.log(results)
       if (err) {
         callback(err);
+      } else if (!results.length) {
+        callback('Access denied');
       } else {
         callback(null, results);
       }
@@ -15,17 +19,52 @@ module.exports = {
   },
   login: function (data, callback) {
     db.connection.connect();
-    db.connection.query(`SELECT * FROM admin WHERE hashword = '${data.hashword}'`, null, (err, results) => {
+    const email = sqlstring.escape(data.email);
+
+    db.connection.query(`SELECT * FROM admin WHERE email = ${email} AND uuid = '${data.uuid}'`, null, (err, results) => {
+      console.log('login verify results: ', results)
       if (err) {
         callback(err);
+      } else if (!results.length) {
+        callback('Access denied');
       } else {
-        callback(null, results);
+        const hashword = results[0].hashword;
+        if (hashword) {
+          bcrypt.compare(data.password, hashword, function(err, isMatch) {
+            if (isMatch) {
+              callback(null, results);
+            } else {
+              callback('Incorrect Password');
+            }
+          });
+        } else {
+          bcrypt.genSalt(10, function(err, salt) {
+            bcrypt.hash(data.password, salt, function(err, hash) {
+              let sessionEnds = new Date();
+              sessionEnds.setMonth(sessionEnds.getMonth() + 2);
+              sessionEnds = sessionEnds.toISOString().split('T')[0];
+              console.log('sessionEnds: ', sessionEnds);
+              const update = `UPDATE admin
+                              SET hashword = '${hash}', session_ends = '${sessionEnds}'
+                              WHERE uuid = '${data.uuid}'`;
+              db.connection.query(update, null, (err, results) => {
+                console.log('update login results: ', results)
+                if (err) {
+                  callback(err);
+                } else {
+                  callback(null, results);
+                }
+              });
+            });
+          });
+        }
       }
     });
   },
-  getAll: function (callback) {
+  getAll: function (data, callback) {
     db.connection.connect();
-    db.connection.query('SELECT * FROM invitees', null, (err, results) => {
+    const getAll = `SELECT * FROM invitees WHERE EXISTS (SELECT * FROM admin WHERE uuid = '${data.uuid}');`
+    db.connection.query(getAll, null, (err, results) => {
       if (err) {
         callback(err);
       } else {
@@ -35,40 +74,37 @@ module.exports = {
   },
   insert: function (data, callback) {
     db.connection.connect();
-
-    db.connection.query(`SELECT * FROM admin WHERE uuid = '${data.uuid}'`, null, (err, results) => {
-      if (err || !results.length) {
+    console.log(data)
+    const name = sqlstring.escape(data.name);
+    const contact = sqlstring.escape(data.contact);
+    const insert_uuid = uuidv4();
+    const insert = `INSERT INTO invitees (uuid, name, contact, guests, language)
+                    SELECT '${insert_uuid}', ${name}, ${contact}, ${data.guests}, ${data.language}
+                    FROM dual
+                    WHERE EXISTS (SELECT * FROM admin WHERE uuid = '${data.uuid}')`;
+    db.connection.query(insert, null, (err, results) => {
+      if (err) {
         callback(err);
       } else {
-        const name = sqlstring.escape(data.name);
-        const contact = sqlstring.escape(data.contact);
-        const insert_uuid = uuidv4();
-        const insert = `INSERT INTO invitees (uuid, name, contact, guests) VALUES ('${insert_uuid}', ${name}, ${contact}, ${data.guests})`;
-        db.connection.query(insert, null, (err, results) => {
+        db.connection.query(`SELECT * FROM invitees`, null, (err, results) => {
           if (err) {
             callback(err);
           } else {
             callback(null, results);
           }
-        });
+        })
       }
     });
   },
   delete: function (data, callback) {
     db.connection.connect();
 
-    db.connection.query(`SELECT * FROM admin WHERE uuid = '${data.uuid}'`, null, (err, results) => {
-      if (err || !results.length) {
+    const deletion = `DELETE FROM invitees WHERE id = ${data.id} AND EXISTS (SELECT * FROM admin WHERE uuid = '${data.uuid}')`;
+    db.connection.query(deletion, null, (err, results) => {
+      if (err) {
         callback(err);
       } else {
-        const deletion = `DELETE FROM invitees WHERE id = ${data.id}`;
-        db.connection.query(deletion, null, (err, results) => {
-          if (err) {
-            callback(err);
-          } else {
-            callback(null, results);
-          }
-        });
+        callback(null, results);
       }
     });
   }
